Added comments and moved add_places function
