+++ app/itinerary/routes.py
@@ -4,7 +4,6 @@ from datetime import timedelta
 
 # INTERNAL
 from ..models import Place, Trip, Day, db, place_schema
-from .helpers import add_places
 from ..global_helpers import create_places_last, serialize_places
 from .itinerary import Itinerary
